[MIG] sale_loyalty_auto_refresh: Migration to 16.0
