--- sale_coupon_auto_refresh/i18n/es.po
@@ -1,99 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* sale_coupon_auto_refresh
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 12.0\n"
-"Report-Msgid-Bugs-To: \n"
-"PO-Revision-Date: 2023-11-08 13:38+0000\n"
-"Last-Translator: Ivorra78 <informatica@totmaterial.es>\n"
-"Language-Team: \n"
-"Language: es\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=2; plural=n != 1;\n"
-"X-Generator: Weblate 4.17\n"
-
-#. module: sale_coupon_auto_refresh
-#: model_terms:ir.ui.view,arch_db:sale_coupon_auto_refresh.res_config_settings_view_form
-msgid ""
-"<span class=\"fa fa-lg fa-building-o\" title=\"Values set here are company-"
-"specific.\" groups=\"base.group_multi_company\"/>"
-msgstr ""
-"<span class=\"fa fa-lg fa-building-o\" title=\"Los valores establecidos aquí "
-"son específicos de la compañía.\" groups=\"base.group_multi_company\"/>"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_sale_coupon_refresh_mixin__auto_refresh_coupon_triggers_data
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_sale_order__auto_refresh_coupon_triggers_data
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_sale_order_line__auto_refresh_coupon_triggers_data
-msgid "Auto Refresh Coupon Triggers Data"
-msgstr "Actualizar Automáticamente los Datos de los Activadores de Cupones"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_res_company__auto_refresh_coupon
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_res_config_settings__auto_refresh_coupon
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_sale_order__auto_refresh_coupon
-msgid "Auto Refresh Coupons"
-msgstr "Auto actualizar ofertas"
-
-#. module: sale_coupon_auto_refresh
-#: model_terms:ir.ui.view,arch_db:sale_coupon_auto_refresh.res_config_settings_view_form
-msgid "Auto refresh coupons"
-msgstr "Auto actualizar ofertas"
-
-#. module: sale_coupon_auto_refresh
-#: model_terms:ir.ui.view,arch_db:sale_coupon_auto_refresh.res_config_settings_view_form
-msgid "Auto-refresh coupons"
-msgstr "Auto actualizar ofertas"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model.fields,help:sale_coupon_auto_refresh.field_res_company__auto_refresh_coupon
-#: model:ir.model.fields,help:sale_coupon_auto_refresh.field_res_config_settings__auto_refresh_coupon
-#: model:ir.model.fields,help:sale_coupon_auto_refresh.field_sale_order__auto_refresh_coupon
-msgid "Autorefresh coupon lines in the backorder"
-msgstr "Auto actualizar las líneas de descuento en backorder"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_res_company
-msgid "Companies"
-msgstr "Compañías"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_res_config_settings
-msgid "Config Settings"
-msgstr "Ajustes"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_sale_coupon_refresh_mixin
-msgid "Mixin class for sale coupon auto-refresh features"
-msgstr "Clase mixta de actualización automática de cupones de descuento"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_sale_coupon_apply_code
-msgid "Sales Coupon Apply Code"
-msgstr "Cupón de descuento Aplicar código"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_sale_order
-msgid "Sales Order"
-msgstr "Orden de Venta"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_sale_order_line
-msgid "Sales Order Line"
-msgstr "Línea de Orden de Venta"
-
-#. module: sale_coupon_auto_refresh
-#: model_terms:ir.ui.view,arch_db:sale_coupon_auto_refresh.res_config_settings_view_form
-msgid ""
-"When the sale order is updated, auto-refresh the promotion programs which "
-"are set to auto apply"
-msgstr ""
-"Cuando se actualice un pedido de venta, auto actualizar los programas de "
-"oferta configurados para auto aplicarse"
-
-#~ msgid "Sale Order"
-#~ msgstr "Pedido de Venta"
