--- sale_coupon_auto_refresh/i18n/zh_CN.po
@@ -1,103 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* sale_coupon_auto_refresh
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"PO-Revision-Date: 2021-09-02 04:50+0000\n"
-"Last-Translator: 黎伟杰 <674416404@qq.com>\n"
-"Language-Team: none\n"
-"Language: zh_CN\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-"X-Generator: Weblate 4.3.2\n"
-
-#. module: sale_coupon_auto_refresh
-#: model_terms:ir.ui.view,arch_db:sale_coupon_auto_refresh.res_config_settings_view_form
-msgid ""
-"<span class=\"fa fa-lg fa-building-o\" title=\"Values set here are company-"
-"specific.\" groups=\"base.group_multi_company\"/>"
-msgstr ""
-"<span class=\"fa fa-lg fa-building-o\" title=\"此处设置的值是特定于公司"
-"的。\" groups=\"base.group_multi_company\"/>"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_sale_coupon_refresh_mixin__auto_refresh_coupon_triggers_data
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_sale_order__auto_refresh_coupon_triggers_data
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_sale_order_line__auto_refresh_coupon_triggers_data
-msgid "Auto Refresh Coupon Triggers Data"
-msgstr ""
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_res_company__auto_refresh_coupon
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_res_config_settings__auto_refresh_coupon
-#: model:ir.model.fields,field_description:sale_coupon_auto_refresh.field_sale_order__auto_refresh_coupon
-msgid "Auto Refresh Coupons"
-msgstr "自动刷新优惠券"
-
-#. module: sale_coupon_auto_refresh
-#: model_terms:ir.ui.view,arch_db:sale_coupon_auto_refresh.res_config_settings_view_form
-msgid "Auto refresh coupons"
-msgstr "自动刷新优惠券"
-
-#. module: sale_coupon_auto_refresh
-#: model_terms:ir.ui.view,arch_db:sale_coupon_auto_refresh.res_config_settings_view_form
-msgid "Auto-refresh coupons"
-msgstr "自动刷新优惠券"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model.fields,help:sale_coupon_auto_refresh.field_res_company__auto_refresh_coupon
-#: model:ir.model.fields,help:sale_coupon_auto_refresh.field_res_config_settings__auto_refresh_coupon
-#: model:ir.model.fields,help:sale_coupon_auto_refresh.field_sale_order__auto_refresh_coupon
-msgid "Autorefresh coupon lines in the backorder"
-msgstr "自动刷新延期交货中的优惠券行"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_res_company
-msgid "Companies"
-msgstr "公司"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_res_config_settings
-msgid "Config Settings"
-msgstr "配置设置"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_sale_coupon_refresh_mixin
-msgid "Mixin class for sale coupon auto-refresh features"
-msgstr ""
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_sale_coupon_apply_code
-msgid "Sales Coupon Apply Code"
-msgstr "销售优惠券申请代码"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_sale_order
-msgid "Sales Order"
-msgstr "销售订单"
-
-#. module: sale_coupon_auto_refresh
-#: model:ir.model,name:sale_coupon_auto_refresh.model_sale_order_line
-msgid "Sales Order Line"
-msgstr "销售订单行"
-
-#. module: sale_coupon_auto_refresh
-#: model_terms:ir.ui.view,arch_db:sale_coupon_auto_refresh.res_config_settings_view_form
-msgid ""
-"When the sale order is updated, auto-refresh the promotion programs which "
-"are set to auto apply"
-msgstr "当销售订单更新时，自动刷新设置为自动应用的促销计划"
-
-#~ msgid "Display Name"
-#~ msgstr "显示名称"
-
-#~ msgid "ID"
-#~ msgstr "ID"
-
-#~ msgid "Last Modified on"
-#~ msgstr "最后修改日"
