--- sale_coupon_auto_refresh/models/__init__.py
@@ -1,4 +1,0 @@
-from . import res_config_settings
-from . import res_company
-from . import sale_coupon_refresh_mixin
-from . import sale_order
