--- sale_coupon_auto_refresh/readme/CONTRIBUTORS.rst
@@ -1,8 +1,0 @@
-* `Tecnativa <https://www.tecnativa.com>`_:
-
-  * David Vidal
-  * Pilar Vargas
-
-* `Camptocamp <https://www.camptocamp.com>`_:
-
-  * Silvio Gregorini
