--- sale_coupon_auto_refresh/readme/USAGE.rst
@@ -1,5 +1,0 @@
-You need to configure a *Promotion program* with the option *Promo Code Usage* set to
-*Automatically applied*. This can be done in *Sales > Products > Promotion programs*.
-
-Now, make a quotation that fulfill the promotion condition and the promotion will be
-automatically applied.
