--- sale_coupon_auto_refresh/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_sale_coupon_auto_refresh
