--- sale_coupon_auto_refresh/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import sale_coupon_apply_code
