+++ sale_loyalty_auto_refresh/__manifest__.py
@@ -1,0 +1,16 @@
+# Copyright 2021 Tecnativa - David Vidal
+# Copyright 2021 Camptocamp - Silvio Gregorini
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+{
+    "name": "Auto Refresh Coupons",
+    "summary": "Allows to auto-apply the coupons with no user intervention",
+    "version": "16.0.1.0.0",
+    "development_status": "Production/Stable",
+    "category": "Website",
+    "website": "https://github.com/OCA/sale-promotion",
+    "author": "Tecnativa, Camptocamp, Odoo Community Association (OCA)",
+    "maintainers": ["chienandalu", "SilvioC2C"],
+    "license": "AGPL-3",
+    "depends": ["sale_loyalty"],
+    "data": ["views/res_config_settings_views.xml"],
+}
