+++ sale_loyalty_auto_refresh/i18n/fr.po
@@ -1,0 +1,105 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* sale_loyalty_auto_refresh
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 14.0\n"
+"Report-Msgid-Bugs-To: \n"
+"PO-Revision-Date: 2021-09-18 14:34+0000\n"
+"Last-Translator: Rémi <remi@le-filament.com>\n"
+"Language-Team: none\n"
+"Language: fr\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: nplurals=2; plural=n > 1;\n"
+"X-Generator: Weblate 4.3.2\n"
+
+#. module: sale_loyalty_auto_refresh
+#: model_terms:ir.ui.view,arch_db:sale_loyalty_auto_refresh.res_config_settings_view_form
+msgid ""
+"<span class=\"fa fa-lg fa-building-o\" title=\"Values set here are company-"
+"specific.\" groups=\"base.group_multi_company\"/>"
+msgstr ""
+"<span class=\"fa fa-lg fa-building-o\" title=\"Ces valeurs sont pour une "
+"seule société.\" groups=\"base.group_multi_company\"/>"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_sale_coupon_refresh_mixin__auto_refresh_coupon_triggers_data
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_sale_order__auto_refresh_coupon_triggers_data
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_sale_order_line__auto_refresh_coupon_triggers_data
+msgid "Auto Refresh Coupon Triggers Data"
+msgstr ""
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_res_company__auto_refresh_coupon
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_res_config_settings__auto_refresh_coupon
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_sale_order__auto_refresh_coupon
+msgid "Auto Refresh Coupons"
+msgstr "Mise à jour automatique des promotions"
+
+#. module: sale_loyalty_auto_refresh
+#: model_terms:ir.ui.view,arch_db:sale_loyalty_auto_refresh.res_config_settings_view_form
+msgid "Auto refresh coupons"
+msgstr "Mise à jour automatique des promotions"
+
+#. module: sale_loyalty_auto_refresh
+#: model_terms:ir.ui.view,arch_db:sale_loyalty_auto_refresh.res_config_settings_view_form
+msgid "Auto-refresh coupons"
+msgstr "Mise à jour automatique des promotions"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model.fields,help:sale_loyalty_auto_refresh.field_res_company__auto_refresh_coupon
+#: model:ir.model.fields,help:sale_loyalty_auto_refresh.field_res_config_settings__auto_refresh_coupon
+#: model:ir.model.fields,help:sale_loyalty_auto_refresh.field_sale_order__auto_refresh_coupon
+msgid "Autorefresh coupon lines in the backorder"
+msgstr "Mise à jour automatique des promotions"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_res_company
+msgid "Companies"
+msgstr "Sociétés"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_res_config_settings
+msgid "Config Settings"
+msgstr "Configuration"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_sale_coupon_refresh_mixin
+msgid "Mixin class for sale coupon auto-refresh features"
+msgstr ""
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_sale_coupon_apply_code
+msgid "Sales Coupon Apply Code"
+msgstr "Utiliser le code promotionnel"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_sale_order
+msgid "Sales Order"
+msgstr "Commandes"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_sale_order_line
+msgid "Sales Order Line"
+msgstr "Ligne de Commande"
+
+#. module: sale_loyalty_auto_refresh
+#: model_terms:ir.ui.view,arch_db:sale_loyalty_auto_refresh.res_config_settings_view_form
+msgid ""
+"When the sale order is updated, auto-refresh the promotion programs which "
+"are set to auto apply"
+msgstr ""
+"Quand le devis est mis à jour, recalcule automatiquement les programmes de "
+"promotions configurés pour application automatique"
+
+#~ msgid "Display Name"
+#~ msgstr "Nom affiché"
+
+#~ msgid "ID"
+#~ msgstr "ID"
+
+#~ msgid "Last Modified on"
+#~ msgstr "Date de dernière modification"
