+++ sale_loyalty_auto_refresh/i18n/it.po
@@ -1,0 +1,105 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* sale_loyalty_auto_refresh
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 14.0\n"
+"Report-Msgid-Bugs-To: \n"
+"PO-Revision-Date: 2022-04-28 10:05+0000\n"
+"Last-Translator: Francesco Foresti <francesco.foresti@ooops404.com>\n"
+"Language-Team: none\n"
+"Language: it\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: nplurals=2; plural=n != 1;\n"
+"X-Generator: Weblate 4.3.2\n"
+
+#. module: sale_loyalty_auto_refresh
+#: model_terms:ir.ui.view,arch_db:sale_loyalty_auto_refresh.res_config_settings_view_form
+msgid ""
+"<span class=\"fa fa-lg fa-building-o\" title=\"Values set here are company-"
+"specific.\" groups=\"base.group_multi_company\"/>"
+msgstr ""
+"<span class=\"fa fa-lg fa-building-o\" title=\"I valori impostati qui sono "
+"specifici per azienda.\" groups=\"base.group_multi_company\"/>"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_sale_coupon_refresh_mixin__auto_refresh_coupon_triggers_data
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_sale_order__auto_refresh_coupon_triggers_data
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_sale_order_line__auto_refresh_coupon_triggers_data
+msgid "Auto Refresh Coupon Triggers Data"
+msgstr "Dati trigger aggiornamento automatic coupons"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_res_company__auto_refresh_coupon
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_res_config_settings__auto_refresh_coupon
+#: model:ir.model.fields,field_description:sale_loyalty_auto_refresh.field_sale_order__auto_refresh_coupon
+msgid "Auto Refresh Coupons"
+msgstr "Aggiornamento automatico coupons"
+
+#. module: sale_loyalty_auto_refresh
+#: model_terms:ir.ui.view,arch_db:sale_loyalty_auto_refresh.res_config_settings_view_form
+msgid "Auto refresh coupons"
+msgstr "Aggiornamento automatico coupons"
+
+#. module: sale_loyalty_auto_refresh
+#: model_terms:ir.ui.view,arch_db:sale_loyalty_auto_refresh.res_config_settings_view_form
+msgid "Auto-refresh coupons"
+msgstr "Aggiornamento automatico coupons"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model.fields,help:sale_loyalty_auto_refresh.field_res_company__auto_refresh_coupon
+#: model:ir.model.fields,help:sale_loyalty_auto_refresh.field_res_config_settings__auto_refresh_coupon
+#: model:ir.model.fields,help:sale_loyalty_auto_refresh.field_sale_order__auto_refresh_coupon
+msgid "Autorefresh coupon lines in the backorder"
+msgstr "Aggiorna automaticamente righe coupon nell'ordine residuo"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_res_company
+msgid "Companies"
+msgstr "Aziende"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_res_config_settings
+msgid "Config Settings"
+msgstr "Impostazioni configurazione"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_sale_coupon_refresh_mixin
+msgid "Mixin class for sale coupon auto-refresh features"
+msgstr "Classe mixin per funzionalità aggiornamento automatico coupon"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_sale_coupon_apply_code
+msgid "Sales Coupon Apply Code"
+msgstr "Applica codice coupon vendite"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_sale_order
+msgid "Sales Order"
+msgstr "Ordine di vendita"
+
+#. module: sale_loyalty_auto_refresh
+#: model:ir.model,name:sale_loyalty_auto_refresh.model_sale_order_line
+msgid "Sales Order Line"
+msgstr "Riga ordine di vendita"
+
+#. module: sale_loyalty_auto_refresh
+#: model_terms:ir.ui.view,arch_db:sale_loyalty_auto_refresh.res_config_settings_view_form
+msgid ""
+"When the sale order is updated, auto-refresh the promotion programs which "
+"are set to auto apply"
+msgstr ""
+"Quando l'ordine di vendita è aggiornato, aggiorna automaticamente le "
+"promozioni che sono impostate ad \"applica automaticamente\""
+
+#~ msgid "Display Name"
+#~ msgstr "Nome Visualizzato"
+
+#~ msgid "ID"
+#~ msgstr "ID"
+
+#~ msgid "Last Modified on"
+#~ msgstr "Ultima modifica il"
