+++ sale_loyalty_auto_refresh/models/__init__.py
@@ -1,0 +1,5 @@
+from . import res_config_settings
+from . import res_company
+from . import sale_coupon_refresh_mixin
+from . import sale_order
+from . import sale_order_line
