+++ sale_loyalty_auto_refresh/models/res_company.py
