+++ sale_loyalty_auto_refresh/models/res_config_settings.py
