+++ sale_loyalty_auto_refresh/readme/CONTRIBUTORS.rst
@@ -1,0 +1,12 @@
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+  * David Vidal
+  * Pilar Vargas
+
+* `Camptocamp <https://www.camptocamp.com>`_:
+
+  * Silvio Gregorini
+
+* `Trobz <https://www.trobz.com>`_:
+
+  * Nguyen Minh Chien <chien@trobz.com>
