+++ sale_loyalty_auto_refresh/readme/DESCRIPTION.rst
