+++ sale_loyalty_auto_refresh/readme/USAGE.rst
@@ -1,0 +1,5 @@
+You need to configure a *Promotion program* with the option *Program trigger* set to
+*Automatic* (shown in Debug mode only, read-only computed field). This can be done in *Sales > Products > Discount & Loyalty*.
+
+Now, make a quotation that fulfill the promotion condition and the promotion will be
+automatically applied.
