+++ sale_loyalty_auto_refresh/static/description/icon.png
