+++ sale_loyalty_auto_refresh/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_sale_loyalty_auto_refresh
