+++ sale_loyalty_auto_refresh/views/res_config_settings_views.xml
