+++ sale_loyalty_auto_refresh/wizard/__init__.py
@@ -1,0 +1,2 @@
+from . import sale_loyalty_coupon_wizard
+from . import sale_loyalty_reward_wizard
