--- setup/sale_coupon_auto_refresh/odoo/addons/sale_coupon_auto_refresh
@@ -1,1 +1,0 @@
-../../../../sale_coupon_auto_refresh
