+++ setup/sale_loyalty_auto_refresh/odoo/addons/sale_loyalty_auto_refresh
@@ -1,0 +1,1 @@
+../../../../sale_loyalty_auto_refresh
